fix tests so they pass with new Substrate updates to commit d03a91a181d0b22d00a6b9ba2a8007dc254779e3

+++ pallets/exchange-rate/src/mock.rs
@@ -44,7 +44,7 @@ parameter_types! {
 }
 impl frame_system::Config for Test {
     type AccountData = pallet_balances::AccountData<u64>;
-    type AccountId = u64;
+	type AccountId = u128; // u64 is not enough to hold bytes used to generate bounty account
     type BaseCallFilter = ();
     type BlockHashCount = BlockHashCount;
     type BlockNumber = u64;
